apps/api/pyproject.toml
[project]
name = "tenue-api"
version = "0.0.0"
description = "Phase 02 FastAPI scaffold for Tenue."
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14,<2.0",
  "fastapi>=0.115,<1.0",
  "httpx>=0.28,<1.0",
  "psycopg[binary]>=3.2,<4.0",
  "python-dotenv>=1.0,<2.0",
  "sqlalchemy>=2.0,<3.0",
  "uvicorn[standard]>=0.34,<1.0",
]

[dependency-groups]
dev = [
  "mypy>=1.15,<2.0",
  "pytest>=8.3,<9.0",
  "ruff>=0.11,<1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "db_integration: requires the local Supabase Postgres instance to be running",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
strict_optional = true
pretty = true
exclude = ["alembic/versions"]
